sentinel-bot/requirements.txt
mysql-connector-python
requests
python-telegram-bot
pymongo